backend/requirements.txt
fastapi
uvicorn
python-dotenv
prisma
python-jose[cryptography]
python-dotenv
passlib[bcrypt]